feat: add ADBC backend
This adds the skeleton of a very minimal ADBC backend. Most tests
do not yet pass. The main thing is that ADBC is really a database
API abstraction akin to PEP 249, JDBC, or ODBC, so this backend
eventually needs to be able to let the user specify the SQL
dialect to compile to (or: to use Substrait instead).